+++ Sources/Main/Core/Private/SceneDelegateHelper.swift
@@ -9,23 +9,21 @@ import RLogger
 
 @available(iOS 13.0, *)
 enum SceneDelegateHelper {
-    private enum Keys {
-        static let aplicationSceneManifestKey = "UIApplicationSceneManifest"
-    }
 
     /// Autotrack the UISceneDelegate functions.
     static func autoTrack() {
-        guard let dict = Bundle.main.object(forInfoDictionaryKey: Keys.aplicationSceneManifestKey) as? [String: Any],
-              let data = try? JSONSerialization.data(withJSONObject: dict, options: []) else {
+        let bundle: EnvironmentBundle = Bundle.main
+
+        guard let applicationSceneManifest = bundle.applicationSceneManifest else {
             RLogger.debug(message: "The app's Info.plist is not configured with UIApplicationSceneManifest.")
             return
         }
-        let decoder = JSONDecoder()
-        let manifest = try? decoder.decode(ApplicationSceneManifest.self, from: data)
-        guard let sceneDelegateClassName = manifest?.sceneConfigurations?.windowSceneSessionRoleApplication?.first?.sceneDelegateClassName else {
+
+        guard let sceneDelegateClassName = applicationSceneManifest.firstSceneDelegateClassName else {
             RLogger.debug(message: "UISceneDelegateClassName could not be retrieved.")
             return
         }
+
         UIWindowScene.rAutotrackSceneDelegateFunctions(sceneDelegateClassName)
         RLogger.debug(message: "\(sceneDelegateClassName) is autotracked.")
     }